dm stub : fix MISRA.ETYPE.COMP.CAST.EXPL.WIDER.2012
MISRA.ETYPE.COMP.CAST.EXPL.WIDER.2012
The value of a composite expression shall not be
cast to a different essential type category or a wider essential type
Type cast to u32.

Fixes: SYSFW-6253

+++ lpm/vim_raw.c
@@ -56,7 +56,7 @@
 
 void vim_set_intr_enable(u32 intr, int enable)
 {
-	u32 mask = 1U << VINTR(intr);
+	u32 mask = ((u32) 1U << VINTR(intr));
 
 	if (enable == INTR_ENABLE) {
 		writel(mask, VIM_BASE + VIM_GRP_EN_SET(intr));
@@ -67,7 +67,7 @@ void vim_set_intr_enable(u32 intr, int enable)
 
 void vim_clear_intr(u32 intr)
 {
-	writel(1U << VINTR(intr), VIM_BASE + VIM_GRP_STS_CLEAR(intr));
+	writel((u32) 1U << VINTR(intr), VIM_BASE + VIM_GRP_STS_CLEAR(intr));
 }
 
 s32 vim_get_intr_number()
